Shiny bindings for TnT

Output and render functions for using TnT within Shiny applications and interactive Rmd documents.

TnTOutput(outputId, width = "100%", height = "400px")

renderTnT(expr, env = parent.frame(), quoted = FALSE)

Arguments

outputId

output variable to read from

width, height

Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.

expr

An expression that generates a TnT

env

The environment in which to evaluate expr.

quoted

Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

Examples

if (interactive() && require(shiny)) {
    ui <- fluidPage(
        fluidRow(TnTOutput("out"))
    )
    server <- function (input, output) {
        output$out <- renderTnT({
            TnT("
              tnt.board.genome()
                .from(0)
                .to(500)
                .min(0)
                .max(1000)
                .add_track(tnt.board.track()
                           .height(20)
                           .color('white')
                           .display(tnt.board.track.feature.axis()))
                .add_track(tnt.board.track()
                           .height(30)
                           .color('#FFCFDD')
                           .data(tnt.board.track.data.sync()
                                 .retriever(function() {return [{start : 200, end : 350}]}))
                           .display(tnt.board.track.feature.block()
                                    .color('blue')
                                    .index(function (d) {return d.start})))
            ")
        })
    }
    shinyApp(ui = ui, server = server)
}